requirements.txt
annotated-types==0.7.0
blinker==1.9.0
blis==1.3.3
catalogue==2.0.10
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
colorama==0.4.6
confection==0.1.5
cryptography==46.0.4
cymem==2.0.13
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
Flask==3.1.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.2
MarkupSafe==3.0.3
murmurhash==1.0.15
numpy==2.4.2
packaging==26.0
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.0
preshed==3.0.12
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pypdfium2==5.4.0
python-docx==1.2.0
requests==2.32.5
smart_open==7.5.0
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
thinc==8.3.10
tqdm==4.67.3
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
wasabi==1.1.3
weasel==0.4.3
Werkzeug==3.1.5
wrapt==2.1.1
